+++ internal/repository/db/postgresql/migrations/000002_create_orders_table.up.sql
@@ -3,7 +3,7 @@ BEGIN TRANSACTION;
 CREATE TABLE orders (
                     order_num VARCHAR UNIQUE NOT NULL,
                     status VARCHAR NOT NULL,
-                    accrual FLOAT,
+                    accrual INT,
                     user_uuid UUID NOT NULL,
                     created_at TIMESTAMP DEFAULT (CURRENT_TIMESTAMP),
                     CONSTRAINT fk_user
